UI Bearing - Vietnam OP Complete
Still requires further work. Idea is to check if the vehicle has a
config entry if it should show bearings or not in it. This way we can
add it to all our existing ones, and not have it accidentally show one
OP.

Alternate idea is to check base class if it's a tank or boat.