podium functions to load operation role gear
On the new map, this will allow for example the medics to interact and
switch to the operation config of a medic, thus being as close to the
experience as possible.

It won't set any special states though, like engineer or medic. Will
retain specalised role sheet slots for the time being.