mortar walk-on function
only available on public currently. artillery is now not firing on the
same position every time. instead, the artillery is caleld to fire upon
the target using any method and the mortar will miss, and learn from
it's past to correct and zero in. General dispersion also increased,
needs testing to see if it's too dispersed now.