CUP AK74 JSRS compat, base target objects for AI to shoot at
marker for previous commit. base target is an invisible object that
needs its ammo set to 0 and set invincible. AI will use all their
heaviest stuff against it and see it at a sensible distance. Mostly for
use in siege missions, ideally I'll be switching the material for
something that can allow bullets through but that would probably be too
easy