L115 reorg, us army tracer switch
gun jesus made an L96 video and it got me in a sniping mood. Putting
l115 and m107 on shotgun recoil profile as a placeholder, 338API is
using RHS API effect when on public and the mag switch to ace bullets
now works on the base class to prevent faffage.