land rover physics update
it's not fast and it's not racy or agile but I'm still pretty happy with
it. Modelled it with locked diffs which means the handbrake works on all
4 wheels and it understeers on dirt above 20mph, but it's better at
bumps and stuff so it's far more applicable to our needs. Also it does
cute little donuts on dirt from a standstill