pub hitpart tweaks
tweak kill threshold so that hollowpoints kill but the much more common
fmj's don't. But a double tap to chest will kill even with normal 9mm
(and buckshot). Puts it in line with us police incidents where the
victim is shot to the ground but still able to pose a threat, then a
single shot to chest makes it final