Fix worldName checker
worldName returns locally loaded map, not mission required map. Check was always true. Now syndicates from the server, what map it's loaded to clients, to check against.

The mission module and init file can seriously be cleaned up though. The mission module is loaded on every client from the mission sqf to begin with, so a lot of these settings don't need to be broadcasted. Some of these functions can be refactored to be locally executed too.