Surgical Kit + UI Changes
ClientPostInit changes. Allows NO HUD as infantry and adds hud when you enter a vehicle. Known flaws are that at the beginning of the op it tends to delete vehicle hud. Can be fixed by entering an exiting vehicle. Possible reason are ACE setting initializing or TWC music intro. A poor fix was adding the hud twice when entering a vehicle. (causes it to flash)