Increased CFGAI skill Floor
This is part of a larger test. My hypothesis is that the randomness of our current ai skill is what is causing situations where and AI murders everyone in a building vs another that never turns around to face us.

I've increase our camouflage in hopes of balancing enemy skill especially at range before we start firing.

I'm very happy with the ace medical setting and ui changes so this is the next thing to tinker with